doc2md/requirements.txt
Markdown
pandas
numpy
Pillow==9.4.0
python-docx
